+++ examples/casEcho.js
@@ -24,11 +24,10 @@
 let restaf   = require('../lib/restaf');
 let payload  = require('./config')('restaf.env');
 let casSetup = require('./lib/casSetup');
-let runAction = require('./lib/runAction');
+// let runAction = require('./lib/runAction');
 let prtUtil  = require('../prtUtil')
 
 let store = restaf.initStore();
-
 async function example() {
        let {session} = await casSetup(store, payload, 'cas');
         
@@ -36,7 +35,7 @@ async function example() {
             action: 'echo',
             data  : {code: 'data casuser.data1; x=1;put x= ; run; '}
         };
-        let r =  await runAction(store, session,p, 'echo');
+        let r =  await store.runAction(session,p);
         return 'done';
     };